Split Configs into files per-subproject
