+++ src/main/scala/test/Configs.scala
@@ -1,0 +1,19 @@
+package testchipip
+
+import chisel3._
+import org.chipsalliance.cde.config.{Parameters, Config}
+
+class WithTestChipUnitTests extends Config((site, here, up) => {
+  case UnitTests => (testParams: Parameters) =>
+    TestChipUnitTests(testParams)
+})
+
+class WithClockUtilTests extends Config((site, here, up) => {
+  case UnitTests => (testParams: Parameters) => ClockUtilTests()
+})
+
+class TestChipUnitTestConfig extends Config(
+  new WithTestChipUnitTests ++ new BaseConfig)
+
+class ClockUtilTestConfig extends Config(
+  new WithClockUtilTests ++ new BaseConfig)
